Aplikacija-kod/server/requirements.txt
apturl                  0.5.2
bcrypt                  3.2.0
beautifulsoup4          4.10.0
blinker                 1.4
Brlapi                  0.8.3
certifi                 2020.6.20
chardet                 4.0.0
chrome-gnome-shell      0.0.0
click                   8.0.3
colorama                0.4.4
command-not-found       0.3
cryptography            3.4.8
cupshelpers             1.0
dbus-python             1.2.18
defer                   1.0.6
distlib                 0.3.6
distro                  1.7.0
distro-info             1.1build1
duplicity               0.8.21
evdev                   1.4.0
fasteners               0.14.1
filelock                3.8.0
Flask                   2.2.2
Flask-JWT-Extended      3.7.0
Flask-Login             0.6.2
future                  0.18.2
gpg                     1.16.0-unknown
html5lib                1.1
httplib2                0.20.2
idna                    3.3
importlib-metadata      4.6.4
itsdangerous            2.1.2
jeepney                 0.7.1
Jinja2                  3.1.2
jsonify                 0.5
keyring                 23.5.0
language-selector       0.1
launchpadlib            1.10.16
lazr.restfulclient      0.14.4
lazr.uri                1.0.6
lockfile                0.12.2
louis                   3.20.0
lxml                    4.8.0
macaroonbakery          1.3.1
Mako                    1.1.3
MarkupSafe              2.1.1
monotonic               1.6
more-itertools          8.10.0
mysql-connector-python  8.0.31
netifaces               0.11.0
oauthlib                3.2.0
olefile                 0.46
paramiko                2.9.3
pexpect                 4.8.0
Pillow                  9.0.1
pip                     22.3.1
platformdirs            2.5.2
protobuf                3.12.4
psutil                  5.9.0
ptyprocess              0.7.0
pycairo                 1.20.1
pycups                  2.0.1
PyGObject               3.42.1
PyJWT                   2.3.0
pymacaroons             0.13.0
PyNaCl                  1.5.0
pyparsing               2.4.7
pyRFC3339               1.1
python-apt              2.3.0+ubuntu2.1
python-dateutil         2.8.1
python-debian           0.1.43ubuntu1
python-magic            0.4.24
python-xlib             0.29
pytz                    2022.1
pyudev                  0.22.0
pyxdg                   0.27
PyYAML                  5.4.1
reportlab               3.6.8
requests                2.25.1
screen-resolution-extra 0.0.0
SecretStorage           3.3.1
setproctitle            1.2.2
setuptools              59.6.0
six                     1.16.0
soupsieve               2.3.1
systemd-python          234
ubuntu-advantage-tools  27.12
ubuntu-drivers-common   0.0.0
ufw                     0.36.1
unattended-upgrades     0.1
urllib3                 1.26.5
usb-creator             0.3.7
vboxapi                 1.0
virtualenv              20.16.5
wadllib                 1.3.6
webencodings            0.5.1
Werkzeug                2.2.2
wheel                   0.37.1
xdg                     5
xkit                    0.0.0
zipp                    1.0.0